api/requirements.txt
flask==3.0.2
flask-cors==4.0.0
firebase-admin==6.2.0
deepgram-sdk==2.12.0
gunicorn==21.2.0
python-dotenv==1.0.1
aiohttp==3.9.3  # For async HTTP requests
sentence-transformers==2.2.2
numpy>=1.24.0
openai>=1.0.0  # For GPT-based chapter generation 